requirements/base.txt
certifi==2019.6.16  
chardet==3.0.4      
coreapi==2.3.3      
coreschema==0.0.4      
Django==2.1        
django-filter==2.1.0      
django-jenkins==0.110.0    
django-rest-swagger==2.2.0      
djangorestframework==3.9.4      
idna==2.8        
itypes==1.1.0      
Jinja2==2.10.1     
MarkupSafe==1.1.1      
openapi-codec==1.3.2      
Pillow==6.0.0           
pytz==2019.1     
requests==2.22.0     
setuptools==41.0.1     
simplejson==3.16.0     
six==1.12.0     
sqlparse==0.3.0      
style==1.1.6      
update==0.0.1      
uritemplate==3.0.0      
urllib3==1.25.3     
